requirements.txt
alembic==1.8.1
amqp==5.1.1
aniso8601==9.0.1
apispec==6.0.0
apispec-webframeworks==0.5.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.6
exceptiongroup==1.0.0
factory-boy==3.2.1
Faker==15.1.3
filelock==3.8.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
importlib-metadata==5.0.0
importlib-resources==5.10.0
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
passlib==1.7.4
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.32
psycopg2==2.9.5
py==1.11.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-factoryboy==2.5.0
pytest-flask==1.2.0
pytest-runner==6.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.42
tomli==2.0.1
typing_extensions==4.4.0
uWSGI==2.0.21
vine==5.0.0
virtualenv==20.16.6
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.10.0
